Populate the form when the user selects Edit
The constraints got all messed up and I don’t have time to fix now.
Need to figure out how to get everything back to normal if the user
doesn’t click the Cancel button if they don’t want to edit anymore.
For example, if they select another item the Delete/Edit menu goes away
but the Save and Cancel buttons and all the information is still on the
form.